refactor(relay): split Handlers god-object into per-domain files
handlers.go was 3,041 lines / 84 functions on one struct spanning ~10
unrelated domains — a navigation and merge-conflict hazard (Architecture
A1 from the audit). Moved the 61 Handle* tool methods into per-domain
files, leaving the struct, constructor and shared helpers (renderTable,
resolveAgent, clampLimit, buildOnboardingPrompt, …) in handlers.go:

  handlers.go            3041 → 762
  handlers_agents.go        (whoami, register, list, deactivate, …)
  handlers_messaging.go     (send, inbox, thread, team inbox, notify)
  handlers_conversations.go
  handlers_memory.go
  handlers_profiles.go
  handlers_tasks.go         (dispatch/claim/.../batch, 15 handlers)
  handlers_boards.go
  handlers_teams.go         (orgs + teams)
  handlers_projects.go
  handlers_locks.go         (dormant file-lock trio)
  handlers_context.go

Pure mechanical move — methods stay on *Handlers, zero behavior change.
go build / go vet clean; full -race suite green.
